+++ wsgidav/default_conf.py
@@ -12,7 +12,7 @@ r"""
     |__/|__/___/\_, /_/____/_/ |_|___/
                /___/
 
-Default confguration.
+Default configuration.
 """
 # from wsgidav.mw.debug_filter import WsgiDavDebugFilter
 from wsgidav.dir_browser import WsgiDavDirBrowser